Revert "fix type"
This reverts commit b5584cc8aabec1b296813c0ba77fc452b21950ed.

+++ libs/compiler/llvmopt.c
@@ -67,7 +67,7 @@ static inline int stack_push_info(information *const info, node_info *const nd)
 
 static inline node_info stack_pop_info(information *const info)
 {
-	node operand_node = node_load(&info->sx->tree, (size_t)stack_pop(&info->nodes));
+	node operand_node = node_load(&info->sx->tree, stack_pop(&info->nodes));
 	size_t operand_depth = (size_t)stack_pop(&info->depths);
 	node_info operand = {&operand_node, operand_depth};
